requirements.txt
Django>=5.1.5
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-phonenumber-field>=7.3.0
pyotp>=2.9.0  # For OTP generation
zxcvbn>=4.4.28  # For password strength validation
